pyproject.toml
[project]
name = "tivit-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.6",
    "orjson>=3.10.13",
    "passlib[bcrypt]>=1.7.4",
    "pyhumps>=3.8.0",
    "pysqlx-engine>=0.3.2",
    "pytest-asyncio>=0.25.1",
    "pytest-cov>=6.0.0",
    "pytest-dotenv>=0.5.2",
    "pytest[standard]>=8.3.4",
    "python-jose[cryptography]>=3.3.0",
    "scalar-fastapi>=1.0.3",
]


# ruff
[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "double"
indent-style = "tab"
docstring-code-format = true
docstring-code-line-length = 120

[tool.ruff.lint]
ignore = ["E402", "F403", "F405"]
